+++ src/main/java/telegramApp/service/TelegramUserService.java
@@ -1,29 +1,12 @@
 package telegramApp.service;
 
-import org.springframework.stereotype.Service;
 import telegramApp.model.TelegramUser;
-import telegramApp.repo.TelegramUserRepo;
 
-import javax.transaction.Transactional;
+public interface TelegramUserService {
 
-@Service
-public class TelegramUserService {
-    private TelegramUserRepo telegramUserRepo;
+    TelegramUser findByChatId(long id);
 
-    public TelegramUserService (TelegramUserRepo telegramUserRepo){
-        this.telegramUserRepo = telegramUserRepo;
-    }
+    void deleteByChatId(Long id);
 
-    @Transactional
-    public TelegramUser findByChatId(long id){
-        return telegramUserRepo.findByChatId(id);
-    }
-
-    public void deleteByChatId (Long id){
-        telegramUserRepo.delete(telegramUserRepo.findByChatId(id));
-    }
-    @Transactional
-    public void addTelegramUser (TelegramUser telegramUser){
-        telegramUserRepo.save(telegramUser);
-    }
+    void addTelegramUser(TelegramUser telegramUser);
 }
